Break Core.scala into bite-sized pieces
